refactor: improve AtomDetail components hierarchy
