+++ src/DevTools/Extension/components/Shell/components/AtomViewer/components/AtomDetail/components/AtomDependentsList.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from 'react';
 import { Box, Code, List, Text } from '@mantine/core';
 import { AnyAtom } from 'src/types';
-import { useAtomsSnapshots } from '../../../../../../../../../hooks/useAtomsSnapshots';
-import { parseDebugLabel } from '../../../../../../../../../utils/parse-debug-label';
+import { useAtomsSnapshots } from '../../../../../../../../hooks/useAtomsSnapshots';
+import { parseDebugLabel } from '../../../../../../../../utils/parse-debug-label';
 
 type AtomDependentsListProps = {
   atom: AnyAtom;
