+++ src/DevTools/Extension/components/Shell/components/AtomViewer/components/AtomDetail/components/AtomParseDeepNestedValue.tsx
@@ -2,9 +2,9 @@ import { memo, useEffect, useState } from 'react';
 import { Box, Text } from '@mantine/core';
 import { useAtomValue } from 'jotai/react';
 import { AnyAtom } from 'src/types';
-import { useUserStore } from '../../../../../../../../../hooks/useUserStore';
-import { AtomValueType } from '../../../../../../../../../utils/get-type-of-atom-value';
-import { deepParseAtomValue } from '../../../../../../../../../utils/parse-atom-value';
+import { useUserStore } from '../../../../../../../../hooks/useUserStore';
+import { AtomValueType } from '../../../../../../../../utils/get-type-of-atom-value';
+import { deepParseAtomValue } from '../../../../../../../../utils/parse-atom-value';
 import {
   MemoizedValueRenderer,
   getPrismLanguageType,
